+++ packages/ai-core/src/common/proxy-util.ts
@@ -84,52 +84,3 @@ export function shouldBypassProxy(targetUrl: string | undefined, noProxyValue: s
     return false;
 }
 
-/**
- * Resolves the proxy URL to use for a given target URL.
- *
- * Resolution order:
- * 1. If `settingsProxy` is provided (e.g., from Theia `http.proxy` preference), use it.
- * 2. Otherwise check environment variables based on the target URL scheme.
- * 3. If a proxy is resolved, check `no_proxy`/`NO_PROXY` — if bypass, return `undefined`.
- * 4. Return the resolved proxy URL or `undefined` if none found.
- *
- * @param targetUrl - The URL for which to resolve a proxy.
- * @param settingsProxy - An optional proxy URL from application settings.
- * @returns The proxy URL to use, or `undefined` if no proxy should be used.
- */
-export function getProxyUrl(targetUrl: string | undefined, settingsProxy?: string): string | undefined {
-    let proxyUrl: string | undefined;
-
-    if (settingsProxy) {
-        proxyUrl = settingsProxy;
-    } else if (targetUrl) {
-        let scheme: string;
-        try {
-            scheme = new URL(targetUrl).protocol;
-        } catch {
-            return undefined;
-        }
-
-        if (scheme === 'https:') {
-            proxyUrl = process.env.https_proxy
-                || process.env.HTTPS_PROXY
-                || process.env.http_proxy
-                || process.env.HTTP_PROXY;
-        } else if (scheme === 'http:') {
-            proxyUrl = process.env.http_proxy
-                || process.env.HTTP_PROXY;
-        }
-    }
-
-    if (!proxyUrl) {
-        return undefined;
-    }
-
-    // Check no_proxy / NO_PROXY
-    const noProxy = process.env.no_proxy ?? process.env.NO_PROXY;
-    if (shouldBypassProxy(targetUrl, noProxy)) {
-        return undefined;
-    }
-
-    return proxyUrl;
-}
