+++ packages/ai-openai/src/browser/openai-frontend-application-contribution.ts
@@ -64,6 +64,7 @@ export class OpenAiFrontendApplicationContribution implements FrontendApplicatio
                     this.updateAllModels();
                 } else if (event.preferenceName === 'http.proxy') {
                     this.manager.setProxyUrl(this.preferenceService.get<string>('http.proxy', undefined));
+                    this.updateAllModels();
                 }
             });
